Rename layout.tsx to Layout.tsx and move to Layout/ directory
