+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'gatsby';
 
-import Layout from '../components/layout';
+import Layout from '../components/Layout/Layout';
 import { SEOTypes } from '../components/SEO';
 import Page from '../components/Layout/Page';
 import Button, { ButtonThemes, ButtonSizes } from '../components/Button';
