+++ src/pages/contact.tsx
@@ -3,7 +3,7 @@ import axios from 'axios';
 import { navigate } from 'gatsby-link';
 
 import { SEOTypes } from '../components/SEO';
-import Layout from '../components/layout';
+import Layout from '../components/Layout/Layout';
 import Form from '../components/contact/form';
 import Page from '../components/Layout/Page';
