+++ src/pages/thanks.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'gatsby';
 
 import { SEOTypes } from '../components/SEO';
-import Layout from '../components/layout';
+import Layout from '../components/Layout/Layout';
 import Page from '../components/Layout/Page';
 import Button, { ButtonThemes, ButtonSizes } from '../components/Button';
