+++ src/templates/blogPost.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { graphql } from 'gatsby';
 import shortid from 'shortid';
 import { SEOTypes } from '../components/SEO';
-import Layout from '../components/layout';
+import Layout from '../components/Layout/Layout';
 import Page from '../components/Layout/Page';
 
 import styles from '../css/BlogPost.module.css';
