+++ src/templates/series.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link, graphql } from 'gatsby';
 import sortBy from 'lodash/sortBy';
-import Layout from '../components/layout';
+import Layout from '../components/Layout/Layout';
 import Page from '../components/Layout/Page';
 
 import styles from '../css/Series.module.css';
